@@ -27,9 +27,9 @@
     "clean:tslint": "rimraf tslint-rules/*.js",
     "compile:tslint": "tsc -P tslint-rules",
     "compile:script": "tsc -P script/tsconfig.json",
-    "lint": "yarn lint:src && yarn lint:prettier",
-    "lint:fix": "yarn lint:src:fix && yarn lint:prettier --fix",
-    "lint:prettier": "ts-node script/prettier.ts",
+    "lint": "yarn prettier && yarn lint:src",
+    "lint:fix": "yarn prettier --write && yarn lint:src:fix",
+    "prettier": "prettier --check \"./**/*.{ts,tsx,js,jsx,scss,html,yaml,yml}\"",
     "lint:src": "yarn tslint && yarn eslint-check && yarn eslint",
     "lint:src:fix": "yarn tslint --fix && yarn eslint --fix",
     "tslint": "tslint -p .",
@@ -41,7 +41,8 @@
     "rebuild-hard:prod": "yarn clean-slate && yarn build:prod",
     "changelog": "ts-node script/changelog/index.ts",
     "draft-release": "ts-node script/draft-release/index.ts",
-    "validate-changelog": "ts-node script/validate-changelog.ts"
+    "validate-changelog": "ts-node script/validate-changelog.ts",
+    "check-modified": "stop-build"
   },
   "author": {
     "name": "GitHub, Inc.",
@@ -98,7 +99,8 @@
     "rimraf": "^2.5.2",
     "sass-loader": "^7.0.1",
     "semver": "^5.5.0",
-    "spectron": "^5.0.0",
+    "spectron": "5.0.0",
+    "stop-build": "^1.1.0",
     "style-loader": "^0.21.0",
     "to-camel-case": "^1.0.0",
     "ts-jest": "^23.1.4",
@@ -162,7 +164,7 @@
     "@types/webpack-merge": "^4.1.3",
     "@types/winston": "^2.2.0",
     "@types/xml2js": "^0.4.0",
-    "electron": "2.0.13",
+    "electron": "3.1.3",
     "electron-builder": "20.28.4",
     "electron-packager": "^12.0.0",
     "electron-winstaller": "2.5.2"
