@@ -16,10 +16,12 @@ export function formatAsLocalRef(name: string): string {
     // In some cases, Git will report this name explicitly to distingush from
     // a remote ref with the same name - this ensures we format it correctly.
     return `refs/${name}`
-  } else {
+  } else if (!name.startsWith('refs/heads/')) {
     // By default Git will drop the heads prefix unless absolutely necessary
     // - include this to ensure the ref is fully qualified.
     return `refs/heads/${name}`
+  } else {
+    return name
   }
 }
 
