@@ -1,4 +1,3 @@
 export { Banner } from './banner'
-export { SuccessfulMerge } from './successful-merge'
-export { MergeConflictsBanner } from './merge-conflicts-banner'
 export { UpdateAvailable } from './update-available'
+export { renderBanner } from './render-banner'
