@@ -23,7 +23,6 @@ import { CSSTransitionGroup } from 'react-transition-group'
 import { openFile } from '../lib/open-file'
 import { Account } from '../../models/account'
 import { PopupType } from '../../models/popup'
-import { enableFileSizeWarningCheck } from '../../lib/feature-flag'
 import { filesNotTrackedByLFS } from '../../lib/git/lfs'
 import { getLargeFilePaths } from '../../lib/large-files'
 import { isConflictedFile, hasUnresolvedConflicts } from '../../lib/status'
@@ -123,43 +122,43 @@ export class ChangesSidebar extends React.Component<IChangesSidebarProps, {}> {
   private onCreateCommit = async (
     context: ICommitContext
   ): Promise<boolean> => {
-    if (enableFileSizeWarningCheck()) {
-      const overSizedFiles = await getLargeFilePaths(
-        this.props.repository,
-        this.props.changes.workingDirectory,
-        100
-      )
-      const filesIgnoredByLFS = await filesNotTrackedByLFS(
-        this.props.repository,
-        overSizedFiles
-      )
+    const { workingDirectory } = this.props.changes
 
-      if (filesIgnoredByLFS.length !== 0) {
-        this.props.dispatcher.showPopup({
-          type: PopupType.OversizedFiles,
-          oversizedFiles: filesIgnoredByLFS,
-          context: context,
-          repository: this.props.repository,
-        })
+    const overSizedFiles = await getLargeFilePaths(
+      this.props.repository,
+      workingDirectory,
+      100
+    )
+    const filesIgnoredByLFS = await filesNotTrackedByLFS(
+      this.props.repository,
+      overSizedFiles
+    )
 
-        return false
-      }
+    if (filesIgnoredByLFS.length !== 0) {
+      this.props.dispatcher.showPopup({
+        type: PopupType.OversizedFiles,
+        oversizedFiles: filesIgnoredByLFS,
+        context: context,
+        repository: this.props.repository,
+      })
+
+      return false
     }
 
     // are any conflicted files left?
-    const conflictedFilesLeft = this.props.changes.workingDirectory.files.filter(
+    const conflictedFilesLeft = workingDirectory.files.filter(
       f =>
         isConflictedFile(f.status) &&
         f.selection.getSelectionType() === DiffSelectionType.None
     )
 
     if (conflictedFilesLeft.length === 0) {
-      this.props.dispatcher.clearMergeConflictsBanner()
+      this.props.dispatcher.clearBanner()
       this.props.dispatcher.recordUnguidedConflictedMergeCompletion()
     }
 
-    // which of the files selected for committing are conflicted?
-    const conflictedFilesSelected = this.props.changes.workingDirectory.files.filter(
+    // which of the files selected for committing are conflicted (with markers)?
+    const conflictedFilesSelected = workingDirectory.files.filter(
       f =>
         isConflictedFile(f.status) &&
         hasUnresolvedConflicts(f.status) &&
@@ -329,8 +328,13 @@ export class ChangesSidebar extends React.Component<IChangesSidebarProps, {}> {
   }
 
   public render() {
-    const changesState = this.props.changes
-    const selectedFileIDs = changesState.selectedFileIDs
+    const {
+      selectedFileIDs,
+      workingDirectory,
+      commitMessage,
+      showCoAuthoredBy,
+      coAuthors,
+    } = this.props.changes
 
     // TODO: I think user will expect the avatar to match that which
     // they have configured in GitHub.com as well as GHE so when we add
@@ -347,7 +351,7 @@ export class ChangesSidebar extends React.Component<IChangesSidebarProps, {}> {
         <ChangesList
           dispatcher={this.props.dispatcher}
           repository={this.props.repository}
-          workingDirectory={changesState.workingDirectory}
+          workingDirectory={workingDirectory}
           selectedFileIDs={selectedFileIDs}
           onFileSelectionChanged={this.onFileSelectionChanged}
           onCreateCommit={this.onCreateCommit}
@@ -363,14 +367,14 @@ export class ChangesSidebar extends React.Component<IChangesSidebarProps, {}> {
           commitAuthor={this.props.commitAuthor}
           branch={this.props.branch}
           gitHubUser={user}
-          commitMessage={this.props.changes.commitMessage}
+          commitMessage={commitMessage}
           focusCommitMessage={this.props.focusCommitMessage}
           autocompletionProviders={this.autocompletionProviders!}
           availableWidth={this.props.availableWidth}
           onIgnore={this.onIgnore}
           isCommitting={this.props.isCommitting}
-          showCoAuthoredBy={this.props.changes.showCoAuthoredBy}
-          coAuthors={this.props.changes.coAuthors}
+          showCoAuthoredBy={showCoAuthoredBy}
+          coAuthors={coAuthors}
           externalEditorLabel={this.props.externalEditorLabel}
           onOpenInExternalEditor={this.props.onOpenInExternalEditor}
           onChangesListScrolled={this.props.onChangesListScrolled}
