@@ -1,34 +1,28 @@
 import * as React from 'react'
-import * as ReactDOM from 'react-dom'
 import { clipboard } from 'electron'
-import { Editor } from 'codemirror'
-import { Disposable } from 'event-kit'
+import { Editor, Doc } from 'codemirror'
 
-import { fatalError } from '../../lib/fatal-error'
-
-import { DiffHunk, DiffLineType, DiffSelection } from '../../models/diff'
+import {
+  DiffHunk,
+  DiffLineType,
+  DiffSelection,
+  DiffLine,
+} from '../../models/diff'
 import {
   WorkingDirectoryFileChange,
   CommittedFileChange,
 } from '../../models/status'
 
-import { Octicon, OcticonSymbol } from '../octicons'
+import { OcticonSymbol } from '../octicons'
 
 import { IEditorConfigurationExtra } from './editor-configuration-extra'
 import { DiffSyntaxMode, IDiffSyntaxModeSpec } from './diff-syntax-mode'
 import { CodeMirrorHost } from './code-mirror-host'
-import { DiffLineGutter } from './diff-line-gutter'
 import {
   diffLineForIndex,
-  diffHunkForIndex,
   findInteractiveDiffRange,
   lineNumberForDiffLine,
 } from './diff-explorer'
-import { RangeSelectionSizePixels } from './edge-detection'
-
-import { ISelectionStrategy } from './selection/selection-strategy'
-import { RangeSelection } from './selection/range-selection-strategy'
-import { DragDropSelection } from './selection/drag-drop-selection-strategy'
 
 import {
   getLineFilters,
@@ -37,6 +31,10 @@ import {
 } from './syntax-highlighting'
 import { relativeChanges } from './changed-range'
 import { Repository } from '../../models/repository'
+import memoizeOne from 'memoize-one'
+import { structuralEquals } from '../../lib/equality'
+import { assertNever } from '../../lib/fatal-error'
+import { clamp } from '../../lib/clamp'
 
 /** The longest line for which we'd try to calculate a line diff. */
 const MaxIntraLineDiffStringLength = 4096
@@ -62,13 +60,62 @@ function highlightParametersEqual(
   newProps: ITextDiffProps,
   prevProps: ITextDiffProps
 ) {
-  if (newProps === prevProps) {
-    return true
-  }
-
   return (
-    newProps.file.id === prevProps.file.id && newProps.text === prevProps.text
+    newProps === prevProps ||
+    (newProps.file.id === prevProps.file.id && newProps.text === prevProps.text)
   )
+}
+
+type SelectionKind = 'hunk' | 'range'
+
+interface ISelection {
+  readonly from: number
+  readonly to: number
+  readonly kind: SelectionKind
+  readonly isSelected: boolean
+}
+
+function createNoNewlineIndicatorWidget() {
+  const widget = document.createElement('span')
+  widget.title = 'No newline at end of file'
+
+  const { w, h, d } = narrowNoNewlineSymbol
+
+  const xmlns = 'http://www.w3.org/2000/svg'
+  const svgElem = document.createElementNS(xmlns, 'svg')
+  svgElem.setAttribute('aria-hidden', 'true')
+  svgElem.setAttribute('version', '1.1')
+  svgElem.setAttribute('viewBox', `0 0 ${w} ${h}`)
+  svgElem.classList.add('no-newline')
+
+  const pathElem = document.createElementNS(xmlns, 'path')
+  pathElem.setAttribute('d', d)
+  pathElem.textContent = 'No newline at end of file'
+  svgElem.appendChild(pathElem)
+
+  widget.appendChild(svgElem)
+  return widget
+}
+
+/**
+ * Utility function to check whether a selection exists, and whether
+ * the given index is contained within the selection.
+ */
+function inSelection(s: ISelection | null, ix: number): s is ISelection {
+  if (s === null) {
+    return false
+  }
+  return ix >= Math.min(s.from, s.to) && ix <= Math.max(s.to, s.from)
+}
+
+/** Utility function for checking whether an event target has a given CSS class */
+function targetHasClass(target: EventTarget | null, token: string) {
+  return target instanceof HTMLElement && target.classList.contains(token)
+}
+
+/** Utility function for checking whether a file supports selection */
+function canSelect(file: ChangedFile): file is WorkingDirectoryFileChange {
+  return file instanceof WorkingDirectoryFileChange
 }
 
 interface ITextDiffProps {
@@ -80,42 +127,93 @@ interface ITextDiffProps {
   readonly hunks: ReadonlyArray<DiffHunk>
 }
 
+const diffGutterName = 'diff-gutter'
+
+const defaultEditorOptions: IEditorConfigurationExtra = {
+  lineNumbers: false,
+  readOnly: true,
+  showCursorWhenSelecting: false,
+  cursorBlinkRate: -1,
+  lineWrapping: true,
+  mode: { name: DiffSyntaxMode.ModeName },
+  // Make sure CodeMirror doesn't capture Tab (and Shift-Tab) and thus destroy tab navigation
+  extraKeys: { Tab: false, 'Shift-Tab': false },
+  scrollbarStyle: __DARWIN__ ? 'simple' : 'native',
+  styleSelectedText: true,
+  lineSeparator: '\n',
+  specialChars: /[\u0000-\u001f\u007f-\u009f\u00ad\u061c\u200b-\u200f\u2028\u2029\ufeff]/,
+  gutters: [diffGutterName],
+}
+
 export class TextDiff extends React.Component<ITextDiffProps, {}> {
   private codeMirror: Editor | null = null
 
-  private gutterWidth: number | null = null
+  private getCodeMirrorDocument = memoizeOne(
+    (text: string, noNewlineIndicatorLines: ReadonlyArray<number>) => {
+      const { mode, firstLineNumber, lineSeparator } = defaultEditorOptions
+      // If the text looks like it could have been formatted using Windows
+      // line endings (\r\n) we need to massage it a bit before we hand it
+      // off to CodeMirror. That's because CodeMirror has two ways of splitting
+      // lines, one is the built in which splits on \n, \r\n and \r. The last
+      // one is important because that will match carriage return characters
+      // inside a diff line. The other way is when consumers supply the
+      // lineSeparator option. That option only takes a string meaning we can
+      // either make it split on '\r\n', '\n' or '\r' but not what we would like
+      // to do, namely '\r?\n'. We want to keep CR characters inside of a diff
+      // line so that we can mark them using the specialChars attribute so
+      // we convert all \r\n to \n and remove any trailing \r character.
+      if (text.indexOf('\r') !== -1) {
+        // Capture the \r if followed by (positive lookahead) a \n or
+        // the end of the string. Note that this does not capture the \n.
+        text = text.replace(/\r(?=\n|$)/g, '')
+      }
+
+      const doc = new Doc(text, mode, firstLineNumber, lineSeparator)
+
+      for (const noNewlineLine of noNewlineIndicatorLines) {
+        doc.setBookmark(
+          { line: noNewlineLine, ch: Infinity },
+          { widget: createNoNewlineIndicatorWidget() }
+        )
+      }
+
+      return doc
+    },
+    // Only re-run the memoization function if the text differs or the array
+    // differs (by structural equality). Allows us to re-use the document as
+    // much as possible, recreating it only if a no-newline appears/disappears
+    structuralEquals
+  )
 
   /**
-   * We store the scroll position before reloading the same diff so that we can
-   * restore it when we're done. If we're not reloading the same diff, this'll
-   * be null.
+   * Returns an array of line numbers that should be marked as lacking a
+   * new line. Memoized such that even if `hunks` changes we don't have
+   * to re-run getCodeMirrorDocument needlessly.
    */
-  private scrollPositionToRestore: { left: number; top: number } | null = null
+  private getNoNewlineIndicatorLines = memoizeOne(
+    (hunks: ReadonlyArray<DiffHunk>) => {
+      const lines = new Array<number>()
+      for (const hunk of hunks) {
+        for (const line of hunk.lines) {
+          if (line.noTrailingNewLine) {
+            lines.push(lineNumberForDiffLine(line, hunks))
+          }
+        }
+      }
+      return lines
+    }
+  )
 
-  /**
-   * A mapping from CodeMirror line handles to disposables which, when disposed
-   * cleans up any line gutter components and events associated with that line.
-   * See renderLine for more information.
-   */
-  private readonly lineCleanup = new Map<any, Disposable>()
+  /** The current, active, diff gutter selection if any */
+  private selection: ISelection | null = null
 
-  /**
-   *  a local cache of gutter elements, keyed by the row in the diff
-   */
-  private cachedGutterElements = new Map<number, DiffLineGutter>()
-
-  /**
-   * Maintain the current state of the user interacting with the diff gutter
-   */
-  private selection: ISelectionStrategy | null = null
+  /** Whether a particular range should be highlighted due to hover */
+  private hunkHighlightRange: ISelection | null = null
 
   private async initDiffSyntaxMode() {
-    const cm = this.codeMirror
-    const file = this.props.file
-    const hunks = this.props.hunks
-    const repo = this.props.repository
+    const { file, hunks, repository } = this.props
 
-    if (!cm) {
+    if (!this.codeMirror) {
       return
     }
 
@@ -125,13 +223,13 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
     const propsSnapshot = this.props
 
     const lineFilters = getLineFilters(hunks)
-    const contents = await getFileContents(repo, file, lineFilters)
+    const contents = await getFileContents(repository, file, lineFilters)
 
     if (!highlightParametersEqual(this.props, propsSnapshot)) {
       return
     }
 
-    const tsOpt = cm.getOption('tabSize')
+    const tsOpt = this.codeMirror.getOption('tabSize')
     const tabSize = typeof tsOpt === 'number' ? tsOpt : 4
 
     const tokens = await highlightContents(contents, tabSize, lineFilters)
@@ -147,389 +245,141 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
       newTokens: tokens.newTokens,
     }
 
-    cm.setOption('mode', spec)
-  }
-
-  private dispose() {
-    this.codeMirror = null
-
-    this.lineCleanup.forEach(disposable => disposable.dispose())
-    this.lineCleanup.clear()
-
-    document.removeEventListener('mouseup', this.onDocumentMouseUp)
+    this.codeMirror.setOption('mode', spec)
   }
 
   /**
-   * compute the diff gutter width based on what's been rendered in the browser
+   * start a selection gesture based on the current interaction
    */
-  private getAndCacheGutterWidth = (): number | null => {
-    if (this.gutterWidth) {
-      return this.gutterWidth
-    }
-
-    if (this.codeMirror) {
-      // as getWidth will return 0 for elements that are offscreen, this code
-      // will look for the first row of the current viewport, which should be
-      // onscreen
-      const viewport = this.codeMirror.getScrollInfo()
-      const top = viewport.top
-
-      const row = this.codeMirror.lineAtHeight(top, 'local')
-      const element = this.cachedGutterElements.get(row)
-
-      if (!element) {
-        console.error(
-          `unable to find element at ${row}, should probably look into that`
-        )
-        return null
-      }
-
-      this.gutterWidth = element.getWidth()
-
-      if (this.gutterWidth === 0) {
-        console.error(
-          `element at row ${row} does not have a width, should probably look into that`
-        )
-      }
-    }
-
-    return this.gutterWidth
-  }
-
-  private updateRangeHoverState = (
-    start: number,
-    end: number,
-    show: boolean
-  ) => {
-    for (let i = start; i <= end; i++) {
-      this.hoverLine(i, show)
-    }
-  }
-
-  private hoverLine = (row: number, include: boolean) => {
-    const element = this.cachedGutterElements.get(row)
-
-    // element may not be drawn by the editor, so updating it isn't necessary
-    if (element) {
-      element.setHover(include)
-    }
-  }
-
-  /**
-   * start a selection gesture based on the current interation
-   */
-  private startSelection = (
+  private startSelection(
     file: WorkingDirectoryFileChange,
     hunks: ReadonlyArray<DiffHunk>,
     index: number,
-    isRangeSelection: boolean
-  ) => {
-    const snapshot = file.selection
-    const selected = snapshot.isSelected(index)
-    const desiredSelection = !selected
+    kind: SelectionKind
+  ) {
+    if (this.selection !== null) {
+      this.cancelSelection()
+    }
 
-    if (isRangeSelection) {
+    const isSelected = !file.selection.isSelected(index)
+
+    if (kind === 'hunk') {
       const range = findInteractiveDiffRange(hunks, index)
       if (!range) {
         console.error('unable to find range for given line in diff')
         return
       }
 
-      this.selection = new RangeSelection(
-        range.start,
-        range.end,
-        desiredSelection,
-        snapshot
-      )
+      const { from, to } = range
+      this.selection = { isSelected, from, to, kind }
+    } else if (kind === 'range') {
+      this.selection = { isSelected, from: index, to: index, kind }
+      document.addEventListener('mousemove', this.onDocumentMouseMove)
     } else {
-      this.selection = new DragDropSelection(index, desiredSelection, snapshot)
+      assertNever(kind, `Unknown selection kind ${kind}`)
     }
 
-    this.selection.paint(this.cachedGutterElements)
-    document.addEventListener('mouseup', this.onDocumentMouseUp)
+    document.addEventListener('mouseup', this.onDocumentMouseUp, { once: true })
   }
 
-  /**
-   * Helper event listener, registered when starting a selection by
-   * clicking anywhere on or near the gutter. Immediately removes itself
-   * from the mouseup event on the document element and ends any current
-   * selection.
-   *
-   * TODO: Once Electron upgrades to Chrome 55 we can drop this in favor
-   * of the 'once' option in addEventListener, see
-   * https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener
-   */
+  private cancelSelection() {
+    if (this.selection) {
+      document.removeEventListener('mouseup', this.onDocumentMouseUp)
+      document.removeEventListener('mousemove', this.onDocumentMouseMove)
+      this.selection = null
+      this.updateViewport()
+    }
+  }
+
+  private onDocumentMouseMove = (ev: MouseEvent) => {
+    if (
+      this.codeMirror === null ||
+      this.selection === null ||
+      this.selection.kind !== 'range'
+    ) {
+      return
+    }
+
+    // CodeMirror can return a line that doesn't exist if the
+    // pointer is placed underneath the last line so we clamp it
+    // to the range of valid values.
+    const max = Math.max(0, this.codeMirror.getDoc().lineCount() - 1)
+    const to = clamp(this.codeMirror.lineAtHeight(ev.y), 0, max)
+
+    this.codeMirror.scrollIntoView({ line: to, ch: 0 })
+
+    if (to !== this.selection.to) {
+      this.selection = { ...this.selection, to }
+      this.updateViewport()
+    }
+  }
+
   private onDocumentMouseUp = (ev: MouseEvent) => {
     ev.preventDefault()
-    document.removeEventListener('mouseup', this.onDocumentMouseUp)
+
+    if (this.selection === null || this.codeMirror === null) {
+      return this.cancelSelection()
+    }
+
+    // A range selection is when the user clicks on the "hunk handle"
+    // which is a hit area spanning 10 or so pixels on either side of
+    // the gutter border, extending into the text area. We capture the
+    // mouse down event on that hunk handle and for the mouse up event
+    // we need to make sure the user is still within that hunk handle
+    // section and in the correct range.
+    if (this.selection.kind === 'hunk') {
+      // Is the pointer over the same range (i.e hunk) that the
+      // selection was originally started from?
+      if (
+        !targetHasClass(ev.target, 'hunk-handle') ||
+        !inSelection(this.selection, this.codeMirror.lineAtHeight(ev.y))
+      ) {
+        return this.cancelSelection()
+      }
+    } else if (this.selection.kind === 'range') {
+      // Special case drag drop selections of 1 as single line 'click'
+      // events for which we require that the cursor is still on the
+      // original gutter element (i.e. if you mouse down on a gutter
+      // element and move the mouse out of the gutter it should not
+      // count as a click when you mouse up)
+      if (this.selection.from === this.selection.to) {
+        if (
+          !targetHasClass(ev.target, 'diff-line-number') &&
+          !targetHasClass(ev.target, 'diff-line-gutter')
+        ) {
+          return this.cancelSelection()
+        }
+      }
+    } else {
+      return assertNever(
+        this.selection.kind,
+        `Unknown selection kind ${this.selection.kind}`
+      )
+    }
+
     this.endSelection()
   }
 
   /**
    * complete the selection gesture and apply the change to the diff
    */
-  private endSelection = () => {
-    if (!this.props.onIncludeChanged || !this.selection) {
-      return
+  private endSelection() {
+    const { onIncludeChanged, file } = this.props
+    if (onIncludeChanged && this.selection && canSelect(file)) {
+      const current = file.selection
+      const { isSelected } = this.selection
+
+      const lower = Math.min(this.selection.from, this.selection.to)
+      const upper = Math.max(this.selection.from, this.selection.to)
+      const length = upper - lower + 1
+
+      onIncludeChanged(current.withRangeSelection(lower, length, isSelected))
+      this.selection = null
     }
-
-    this.props.onIncludeChanged(this.selection.done())
-
-    // operation is completed, clean this up
-    this.selection = null
-  }
-
-  private onGutterMouseDown = (
-    index: number,
-    hunks: ReadonlyArray<DiffHunk>,
-    isRangeSelection: boolean
-  ) => {
-    if (!(this.props.file instanceof WorkingDirectoryFileChange)) {
-      fatalError(
-        'must not start selection when selected file is not a WorkingDirectoryFileChange'
-      )
-      return
-    }
-
-    if (isRangeSelection) {
-      const hunk = diffHunkForIndex(hunks, index)
-      if (!hunk) {
-        console.error('unable to find hunk for given line in diff')
-      }
-    }
-    this.startSelection(this.props.file, hunks, index, isRangeSelection)
-  }
-
-  private onGutterMouseMove = (index: number) => {
-    if (!this.selection) {
-      return
-    }
-
-    this.selection.update(index)
-    this.selection.paint(this.cachedGutterElements)
-  }
-
-  private onDiffTextMouseMove = (
-    ev: MouseEvent,
-    hunks: ReadonlyArray<DiffHunk>,
-    index: number
-  ) => {
-    const isActive = this.isMouseCursorNearGutter(ev)
-    if (isActive === null) {
-      return
-    }
-
-    const diffLine = diffLineForIndex(hunks, index)
-    if (!diffLine) {
-      return
-    }
-
-    if (!diffLine.isIncludeableLine()) {
-      return
-    }
-
-    const range = findInteractiveDiffRange(hunks, index)
-    if (!range) {
-      console.error('unable to find range for given index in diff')
-      return
-    }
-
-    this.updateRangeHoverState(range.start, range.end, isActive)
-  }
-
-  private onDiffTextMouseDown = (
-    ev: MouseEvent,
-    hunks: ReadonlyArray<DiffHunk>,
-    index: number
-  ) => {
-    const isActive = this.isMouseCursorNearGutter(ev)
-
-    if (isActive) {
-      // this line is important because it prevents the codemirror editor
-      // from handling the event and resetting the scroll position.
-      // it doesn't do this when you click on elements in the gutter,
-      // which is an amazing joke to have placed upon me right now
-      ev.preventDefault()
-
-      if (!(this.props.file instanceof WorkingDirectoryFileChange)) {
-        fatalError(
-          'must not start selection when selected file is not a WorkingDirectoryFileChange'
-        )
-        return
-      }
-
-      this.startSelection(this.props.file, hunks, index, true)
-    }
-  }
-
-  private onDiffTextMouseLeave = (
-    ev: MouseEvent,
-    hunks: ReadonlyArray<DiffHunk>,
-    index: number
-  ) => {
-    const range = findInteractiveDiffRange(hunks, index)
-    if (!range) {
-      console.error('unable to find range for given index in diff')
-      return
-    }
-
-    this.updateRangeHoverState(range.start, range.end, false)
-  }
-
-  private isMouseCursorNearGutter = (ev: MouseEvent): boolean | null => {
-    const width = this.getAndCacheGutterWidth()
-
-    if (!width) {
-      // should fail earlier than this with a helpful error message
-      return null
-    }
-
-    const deltaX = ev.layerX - width
-    return deltaX >= 0 && deltaX <= RangeSelectionSizePixels
   }
 
   private isSelectionEnabled = () => {
-    return this.selection == null
-  }
-
-  private restoreScrollPosition(cm: Editor) {
-    const scrollPosition = this.scrollPositionToRestore
-    if (cm && scrollPosition) {
-      cm.scrollTo(scrollPosition.left, scrollPosition.top)
-    }
-  }
-
-  private renderLine = (instance: any, line: any, element: HTMLElement) => {
-    const existingLineDisposable = this.lineCleanup.get(line)
-
-    // If we can find the line in our cleanup list that means the line is
-    // being re-rendered. Agains, CodeMirror doesn't fire the 'delete' event
-    // when this happens.
-    if (existingLineDisposable) {
-      existingLineDisposable.dispose()
-      this.lineCleanup.delete(line)
-    }
-
-    const index = instance.getLineNumber(line) as number
-
-    const diffLine = diffLineForIndex(this.props.hunks, index)
-    if (diffLine) {
-      const diffLineElement = element.children[0] as HTMLSpanElement
-
-      let noNewlineReactContainer: HTMLSpanElement | null = null
-
-      if (diffLine.noTrailingNewLine) {
-        noNewlineReactContainer = document.createElement('span')
-        noNewlineReactContainer.setAttribute(
-          'title',
-          'No newline at end of file'
-        )
-        ReactDOM.render(
-          <Octicon symbol={narrowNoNewlineSymbol} className="no-newline" />,
-          noNewlineReactContainer
-        )
-        diffLineElement.appendChild(noNewlineReactContainer)
-      }
-
-      const gutterReactContainer = document.createElement('span')
-
-      let isIncluded = false
-      if (this.props.file instanceof WorkingDirectoryFileChange) {
-        isIncluded = this.props.file.selection.isSelected(index)
-      }
-
-      const cache = this.cachedGutterElements
-
-      const hunks = this.props.hunks
-
-      ReactDOM.render(
-        <DiffLineGutter
-          line={diffLine}
-          isIncluded={isIncluded}
-          index={index}
-          readOnly={this.props.readOnly}
-          hunks={hunks}
-          updateRangeHoverState={this.updateRangeHoverState}
-          isSelectionEnabled={this.isSelectionEnabled}
-          onMouseDown={this.onGutterMouseDown}
-          onMouseMove={this.onGutterMouseMove}
-        />,
-        gutterReactContainer,
-        function(this: DiffLineGutter) {
-          if (this !== undefined) {
-            cache.set(index, this)
-          }
-        }
-      )
-
-      const onMouseMoveLine: (ev: MouseEvent) => void = ev => {
-        this.onDiffTextMouseMove(ev, hunks, index)
-      }
-
-      const onMouseDownLine: (ev: MouseEvent) => void = ev => {
-        this.onDiffTextMouseDown(ev, hunks, index)
-      }
-
-      const onMouseLeaveLine: (ev: MouseEvent) => void = ev => {
-        this.onDiffTextMouseLeave(ev, hunks, index)
-      }
-
-      if (!this.props.readOnly) {
-        diffLineElement.addEventListener('mousemove', onMouseMoveLine)
-        diffLineElement.addEventListener('mousedown', onMouseDownLine)
-        diffLineElement.addEventListener('mouseleave', onMouseLeaveLine)
-      }
-
-      element.insertBefore(gutterReactContainer, diffLineElement)
-
-      // Hack(ish?). In order to be a real good citizen we need to unsubscribe from
-      // the line delete event once we've been called once or the component has been
-      // unmounted. In the latter case it's _probably_ not strictly necessary since
-      // the only thing gc rooted by the event should be isolated and eligble for
-      // collection. But let's be extra cautious I guess.
-      //
-      // The only way to unsubscribe is to pass the exact same function given to the
-      // 'on' function to the 'off' so we need a reference to ourselves, basically.
-      let deleteHandler: () => void // eslint-disable-line prefer-const
-
-      // Since we manually render a react component we have to take care of unmounting
-      // it or else we'll leak memory. This disposable will unmount the component.
-      //
-      // See https://facebook.github.io/react/blog/2015/10/01/react-render-and-top-level-api.html
-      const gutterCleanup = new Disposable(() => {
-        this.cachedGutterElements.delete(index)
-
-        ReactDOM.unmountComponentAtNode(gutterReactContainer)
-
-        if (noNewlineReactContainer) {
-          ReactDOM.unmountComponentAtNode(noNewlineReactContainer)
-        }
-
-        if (!this.props.readOnly) {
-          diffLineElement.removeEventListener('mousemove', onMouseMoveLine)
-          diffLineElement.removeEventListener('mousedown', onMouseDownLine)
-          diffLineElement.removeEventListener('mouseleave', onMouseLeaveLine)
-        }
-
-        line.off('delete', deleteHandler)
-      })
-
-      // Add the cleanup disposable to our list of disposables so that we clean up when
-      // this component is unmounted or when the line is re-rendered. When either of that
-      // happens the line 'delete' event doesn't  fire.
-      this.lineCleanup.set(line, gutterCleanup)
-
-      // If the line delete event fires we dispose of the disposable (disposing is
-      // idempotent)
-      deleteHandler = () => {
-        const disp = this.lineCleanup.get(line)
-        if (disp) {
-          this.lineCleanup.delete(line)
-          disp.dispose()
-        }
-      }
-      line.on('delete', deleteHandler)
-    }
+    return this.selection === null
   }
 
   private getAndStoreCodeMirrorInstance = (cmh: CodeMirrorHost | null) => {
@@ -565,10 +415,7 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
     }
   }
 
-  private markIntraLineChanges(
-    codeMirror: Editor,
-    hunks: ReadonlyArray<DiffHunk>
-  ) {
+  private markIntraLineChanges(doc: Doc, hunks: ReadonlyArray<DiffHunk>) {
     for (const hunk of hunks) {
       const additions = hunk.lines.filter(l => l.type === DiffLineType.Add)
       const deletions = hunk.lines.filter(l => l.type === DiffLineType.Delete)
@@ -602,9 +449,7 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
               line: addLineNumber,
               ch: addRange.location + addRange.length + 1,
             }
-            codeMirror
-              .getDoc()
-              .markText(addFrom, addTo, { className: 'cm-diff-add-inner' })
+            doc.markText(addFrom, addTo, { className: 'cm-diff-add-inner' })
           }
         }
 
@@ -620,7 +465,7 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
               line: deleteLineNumber,
               ch: deleteRange.location + deleteRange.length + 1,
             }
-            codeMirror.getDoc().markText(deleteFrom, deleteTo, {
+            doc.markText(deleteFrom, deleteTo, {
               className: 'cm-diff-delete-inner',
             })
           }
@@ -629,85 +474,265 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
     }
   }
 
-  private onChanges = (cm: Editor) => {
-    this.restoreScrollPosition(cm)
-
-    this.markIntraLineChanges(cm, this.props.hunks)
+  private onSwapDoc = (cm: Editor, oldDoc: Doc) => {
+    this.initDiffSyntaxMode()
+    this.markIntraLineChanges(cm.getDoc(), this.props.hunks)
   }
 
-  public componentWillReceiveProps(nextProps: ITextDiffProps) {
-    // If we're reloading the same file, we want to save the current scroll
-    // position and restore it after the diff's been updated.
-    const sameFile =
-      nextProps.file &&
-      this.props.file &&
-      nextProps.file.id === this.props.file.id
+  private onViewportChange = (cm: Editor, from: number, to: number) => {
+    const doc = cm.getDoc()
+    const batchedOps = new Array<Function>()
 
-    // Happy path, if the text hasn't changed we won't re-render
-    // and subsequently won't have to restore the scroll position.
-    const textHasChanged = nextProps.text !== this.props.text
+    doc.eachLine(from, to, line => {
+      const lineNumber = doc.getLineNumber(line)
 
-    const codeMirror = this.codeMirror
-    if (codeMirror && sameFile && textHasChanged) {
-      const scrollInfo = codeMirror.getScrollInfo()
-      this.scrollPositionToRestore = {
-        left: scrollInfo.left,
-        top: scrollInfo.top,
-      }
-    } else {
-      this.scrollPositionToRestore = null
-    }
+      if (lineNumber !== null) {
+        const diffLine = diffLineForIndex(this.props.hunks, lineNumber)
 
-    if (codeMirror && this.props.text !== nextProps.text) {
-      codeMirror.setOption('mode', { name: DiffSyntaxMode.ModeName })
-    }
+        if (diffLine !== null) {
+          const lineInfo = cm.lineInfo(line)
 
-    // HACK: This entire section is a hack. Whenever we receive
-    // props we update all currently visible gutter elements with
-    // the selection state from the file.
-    if (nextProps.file instanceof WorkingDirectoryFileChange) {
-      const selection = nextProps.file.selection
-      const oldSelection =
-        this.props.file instanceof WorkingDirectoryFileChange
-          ? this.props.file.selection
-          : null
-
-      // Nothing has changed
-      if (oldSelection === selection) {
-        return
-      }
-
-      this.gutterWidth = null
-
-      const hunks = nextProps.hunks
-      this.cachedGutterElements.forEach((element, index) => {
-        if (!element) {
-          console.error('expected DOM element for diff gutter not found')
-          return
+          if (
+            lineInfo.gutterMarkers &&
+            diffGutterName in lineInfo.gutterMarkers
+          ) {
+            const marker = lineInfo.gutterMarkers[diffGutterName]
+            if (marker instanceof HTMLElement) {
+              this.updateGutterMarker(marker, lineNumber, diffLine)
+            }
+          } else {
+            batchedOps.push(() => {
+              const marker = this.createGutterMarker(lineNumber, diffLine)
+              cm.setGutterMarker(line, diffGutterName, marker)
+            })
+          }
         }
+      }
+    })
 
-        const line = diffLineForIndex(hunks, index)
-        const isIncludable = line ? line.isIncludeableLine() : false
-        const isSelected = selection.isSelected(index) && isIncludable
-        element.setSelected(isSelected)
-      })
+    // Updating a gutter marker doesn't affect layout or rendering
+    // as far as CodeMirror is concerned so we only run an operation
+    // (which will trigger a CodeMirror refresh) when we have gutter
+    // markers to create.
+    if (batchedOps.length > 0) {
+      cm.operation(() => batchedOps.forEach(x => x()))
     }
   }
 
-  public componentWillUnmount() {
-    this.dispose()
+  /**
+   * Returns a value indicating whether the given line index is included
+   * in the current temporary or permanent (props) selection. Note that
+   * this function does not care about whether the line can be included,
+   * only whether it is indicated to be included by either selection.
+   */
+  private isIncluded(index: number) {
+    const { file } = this.props
+    return inSelection(this.selection, index)
+      ? this.selection.isSelected
+      : canSelect(file) && file.selection.isSelected(index)
   }
 
-  public componentDidUpdate(prevProps: ITextDiffProps) {
-    if (this.props.text === prevProps.text) {
+  private getGutterLineClassNameInfo(
+    index: number,
+    diffLine: DiffLine
+  ): { [className: string]: boolean } {
+    const isIncludeable = diffLine.isIncludeableLine()
+    const isIncluded = isIncludeable && this.isIncluded(index)
+    const hover = isIncludeable && inSelection(this.hunkHighlightRange, index)
+
+    return {
+      'diff-line-gutter': true,
+      'diff-add': diffLine.type === DiffLineType.Add,
+      'diff-delete': diffLine.type === DiffLineType.Delete,
+      'diff-context': diffLine.type === DiffLineType.Context,
+      'diff-hunk': diffLine.type === DiffLineType.Hunk,
+      'read-only': this.props.readOnly,
+      'diff-line-selected': isIncluded,
+      'diff-line-hover': hover,
+      includeable: isIncludeable && !this.props.readOnly,
+    }
+  }
+
+  private createGutterMarker(index: number, diffLine: DiffLine) {
+    const marker = document.createElement('div')
+    marker.className = 'diff-line-gutter'
+
+    marker.addEventListener('mousedown', this.onDiffLineGutterMouseDown)
+
+    const oldLineNumber = document.createElement('div')
+    oldLineNumber.classList.add('diff-line-number', 'before')
+    marker.appendChild(oldLineNumber)
+
+    const newLineNumber = document.createElement('div')
+    newLineNumber.classList.add('diff-line-number', 'after')
+    marker.appendChild(newLineNumber)
+
+    const hunkHandle = document.createElement('div')
+    hunkHandle.addEventListener('mouseenter', this.onHunkHandleMouseEnter)
+    hunkHandle.addEventListener('mouseleave', this.onHunkHandleMouseLeave)
+    hunkHandle.addEventListener('mousedown', this.onHunkHandleMouseDown)
+    hunkHandle.classList.add('hunk-handle')
+    marker.appendChild(hunkHandle)
+
+    this.updateGutterMarker(marker, index, diffLine)
+
+    return marker
+  }
+
+  private updateGutterMarker(
+    marker: HTMLElement,
+    index: number,
+    diffLine: DiffLine
+  ) {
+    const classNameInfo = this.getGutterLineClassNameInfo(index, diffLine)
+    for (const [className, include] of Object.entries(classNameInfo)) {
+      if (include) {
+        marker.classList.add(className)
+      } else {
+        marker.classList.remove(className)
+      }
+    }
+
+    if (!this.props.readOnly && diffLine.isIncludeableLine()) {
+      marker.setAttribute('role', 'button')
+    } else {
+      marker.removeAttribute('role')
+    }
+
+    const oldLineNumber = marker.childNodes[0]
+
+    oldLineNumber.textContent =
+      diffLine.oldLineNumber === null ? '' : `${diffLine.oldLineNumber}`
+
+    const newLineNumber = marker.childNodes[1]
+
+    newLineNumber.textContent =
+      diffLine.newLineNumber === null ? '' : `${diffLine.newLineNumber}`
+  }
+
+  private onHunkHandleMouseEnter = (ev: MouseEvent) => {
+    if (
+      this.codeMirror === null ||
+      this.props.readOnly ||
+      (this.selection !== null && this.selection.kind === 'range')
+    ) {
+      return
+    }
+    const lineNumber = this.codeMirror.lineAtHeight(ev.y)
+
+    const diffLine = diffLineForIndex(this.props.hunks, lineNumber)
+
+    if (!diffLine || !diffLine.isIncludeableLine()) {
       return
     }
 
-    if (this.codeMirror) {
-      this.codeMirror.setOption('mode', { name: DiffSyntaxMode.ModeName })
+    const range = findInteractiveDiffRange(this.props.hunks, lineNumber)
+
+    if (range === null) {
+      return
     }
 
-    this.initDiffSyntaxMode()
+    const { from, to } = range
+    this.hunkHighlightRange = { from, to, kind: 'hunk', isSelected: false }
+    this.updateViewport()
+  }
+
+  private updateViewport() {
+    if (this.codeMirror) {
+      const { from, to } = this.codeMirror.getViewport()
+      this.onViewportChange(this.codeMirror, from, to)
+    }
+  }
+
+  private onDiffLineGutterMouseDown = (ev: MouseEvent) => {
+    // If the event is prevented that means the hunk handle was
+    // clicked first and prevented the default action so we'll bail.
+    if (ev.defaultPrevented || this.codeMirror === null) {
+      return
+    }
+
+    const { file, hunks, readOnly } = this.props
+
+    if (!canSelect(file) || readOnly) {
+      return
+    }
+
+    ev.preventDefault()
+
+    const lineNumber = this.codeMirror.lineAtHeight(ev.y)
+    this.startSelection(file, hunks, lineNumber, 'range')
+  }
+
+  private onHunkHandleMouseLeave = (ev: MouseEvent) => {
+    if (this.hunkHighlightRange !== null) {
+      this.hunkHighlightRange = null
+      this.updateViewport()
+    }
+  }
+
+  private onHunkHandleMouseDown = (ev: MouseEvent) => {
+    if (!this.codeMirror) {
+      return
+    }
+
+    const { file, hunks, readOnly } = this.props
+
+    if (!canSelect(file) || readOnly) {
+      return
+    }
+
+    ev.preventDefault()
+
+    const lineNumber = this.codeMirror.lineAtHeight(ev.y)
+    this.startSelection(file, hunks, lineNumber, 'hunk')
+  }
+
+  public componentWillUnmount() {
+    this.cancelSelection()
+    this.codeMirror = null
+  }
+
+  public componentDidUpdate(
+    prevProps: ITextDiffProps,
+    prevState: {},
+    // tslint:disable-next-line:react-proper-lifecycle-methods
+    snapshot: CodeMirror.ScrollInfo | null
+  ) {
+    if (this.codeMirror === null) {
+      return
+    }
+
+    if (canSelect(this.props.file)) {
+      if (
+        !canSelect(prevProps.file) ||
+        this.props.file.selection !== prevProps.file.selection
+      ) {
+        // If the text has changed the gutters will be recreated
+        // regardless but if it hasn't then we'll need to update
+        // the viewport.
+        if (this.props.text === prevProps.text) {
+          this.updateViewport()
+        }
+      }
+    }
+
+    if (snapshot !== null) {
+      this.codeMirror.scrollTo(undefined, snapshot.top)
+    }
+  }
+
+  public getSnapshotBeforeUpdate(prevProps: ITextDiffProps) {
+    // Store the scroll position when the file stays the same
+    // but we probably swapped out the document
+    if (
+      this.codeMirror !== null &&
+      this.props.file !== prevProps.file &&
+      this.props.file.id === prevProps.file.id &&
+      this.props.text !== prevProps.text
+    ) {
+      return this.codeMirror.getScrollInfo()
+    }
+    return null
   }
 
   public componentDidMount() {
@@ -715,48 +740,19 @@ export class TextDiff extends React.Component<ITextDiffProps, {}> {
   }
 
   public render() {
-    const options: IEditorConfigurationExtra = {
-      lineNumbers: false,
-      readOnly: true,
-      showCursorWhenSelecting: false,
-      cursorBlinkRate: -1,
-      lineWrapping: true,
-      mode: { name: DiffSyntaxMode.ModeName },
-      // Make sure CodeMirror doesn't capture Tab (and Shift-Tab) and thus destroy tab navigation
-      extraKeys: { Tab: false, 'Shift-Tab': false },
-      scrollbarStyle: __DARWIN__ ? 'simple' : 'native',
-      styleSelectedText: true,
-      lineSeparator: '\n',
-      specialChars: /[\u0000-\u001f\u007f-\u009f\u00ad\u061c\u200b-\u200f\u2028\u2029\ufeff]/,
-    }
-
-    // If the text looks like it could have been formatted using Windows
-    // line endings (\r\n) we need to massage it a bit before we hand it
-    // off to CodeMirror. That's because CodeMirror has two ways of splitting
-    // lines, one is the built in which splits on \n, \r\n and \r. The last
-    // one is important because that will match carriage return characters
-    // inside a diff line. The other way is when consumers supply the
-    // lineSeparator option. That option only takes a string meaning we can
-    // either make it split on '\r\n', '\n' or '\r' but not what we would like
-    // to do, namely '\r?\n'. We want to keep CR characters inside of a diff
-    // line so that we can mark them using the specialChars attribute so
-    // we convert all \r\n to \n and remove any trailing \r character.
-    const text =
-      this.props.text.indexOf('\r') !== -1
-        ? this.props.text
-            // Capture the \r if followed by (positive lookahead) a \n or
-            // the end of the string. Note that this does not capture the \n.
-            .replace(/\r(?=\n|$)/g, '')
-        : this.props.text
+    const doc = this.getCodeMirrorDocument(
+      this.props.text,
+      this.getNoNewlineIndicatorLines(this.props.hunks)
+    )
 
     return (
       <CodeMirrorHost
         className="diff-code-mirror"
-        value={text}
-        options={options}
+        value={doc}
+        options={defaultEditorOptions}
         isSelectionEnabled={this.isSelectionEnabled}
-        onChanges={this.onChanges}
-        onRenderLine={this.renderLine}
+        onSwapDoc={this.onSwapDoc}
+        onViewportChange={this.onViewportChange}
         ref={this.getAndStoreCodeMirrorInstance}
         onCopy={this.onCopy}
       />
