@@ -40,7 +40,7 @@ export class CommitConflictsWarning extends React.Component<
       this.props.repository,
       this.props.context
     )
-    this.props.dispatcher.clearMergeConflictsBanner()
+    this.props.dispatcher.clearBanner()
     this.props.dispatcher.setCommitMessage(
       this.props.repository,
       DefaultCommitMessage
