@@ -1,5 +1,4 @@
 import * as React from 'react'
-import { join } from 'path'
 import { Button } from '../lib/button'
 import { ButtonGroup } from '../lib/button-group'
 import { Dialog, DialogContent, DialogFooter } from '../dialog'
@@ -10,25 +9,20 @@ import { Repository } from '../../models/repository'
 import {
   WorkingDirectoryStatus,
   WorkingDirectoryFileChange,
-  AppFileStatusKind,
-  ConflictedFileStatus,
-  isConflictedFileStatus,
-  isConflictWithMarkers,
 } from '../../models/status'
 import { Octicon, OcticonSymbol } from '../octicons'
-import { PathText } from '../lib/path-text'
 import { DialogHeader } from '../dialog/header'
 import { LinkButton } from '../lib/link-button'
-import { isConflictedFile, hasUnresolvedConflicts } from '../../lib/status'
-import { DefaultCommitMessage } from '../../models/commit-message'
-import { shell } from '../../lib/app-shell'
-import { openFile } from '../lib/open-file'
-import { showContextualMenu } from '../main-process-proxy'
-import { IMenuItem } from '../../lib/menu-item'
 import {
-  OpenWithDefaultProgramLabel,
-  RevealInFileManagerLabel,
-} from '../lib/context-menu'
+  isConflictedFile,
+  getResolvedFiles,
+  getConflictedFiles,
+  getUnmergedFiles,
+} from '../../lib/status'
+import { DefaultCommitMessage } from '../../models/commit-message'
+import { renderUnmergedFile } from './unmerged-file'
+import { ManualConflictResolution } from '../../models/manual-conflict-resolution'
+import { BannerType } from '../../models/banner'
 
 interface IMergeConflictsDialogProps {
   readonly dispatcher: Dispatcher
@@ -41,53 +35,7 @@ interface IMergeConflictsDialogProps {
   readonly ourBranch: string
   /* `undefined` when we didn't know the branch at the beginning of this flow */
   readonly theirBranch?: string
-}
-
-/**
- * Calculates the number of merge conclicts in a file from the number of markers
- * divides by three and rounds up since each conflict is indicated by three separate markers
- * (`<<<<<`, `>>>>>`, and `=====`)
- * @param conflictMarkers number of conflict markers in a file
- */
-function calculateConflicts(conflictMarkers: number) {
-  return Math.ceil(conflictMarkers / 3)
-}
-
-/** Filter working directory changes for conflicted or resolved files  */
-function getUnmergedFiles(status: WorkingDirectoryStatus) {
-  return status.files.filter(f => isConflictedFile(f.status))
-}
-
-/** Filter working directory changes for resolved files  */
-function getResolvedFiles(status: WorkingDirectoryStatus) {
-  return status.files.filter(
-    f => isConflictedFileStatus(f.status) && !hasUnresolvedConflicts(f.status)
-  )
-}
-
-/** Filter working directory changes for conflicted files  */
-function getConflictedFiles(status: WorkingDirectoryStatus) {
-  return status.files.filter(
-    f => isConflictedFileStatus(f.status) && hasUnresolvedConflicts(f.status)
-  )
-}
-
-function editorButtonString(editorName: string | null): string {
-  const defaultEditorString = 'editor'
-  return `Open in ${editorName || defaultEditorString}`
-}
-
-function editorButtonTooltip(editorName: string | null): string | undefined {
-  if (editorName !== null) {
-    // no need to render a tooltip if we have a known editor
-    return
-  }
-
-  if (__DARWIN__) {
-    return `No editor configured in Preferences > Advanced`
-  } else {
-    return `No editor configured in Options > Advanced`
-  }
+  readonly manualResolutions: Map<string, ManualConflictResolution>
 }
 
 const submitButtonString = 'Commit merge'
@@ -112,6 +60,7 @@ export class MergeConflictsDialog extends React.Component<
       this.props.repository,
       this.props.workingDirectory,
       {
+        type: BannerType.SuccessfulMerge,
         ourBranch: this.props.ourBranch,
         theirBranch: this.props.theirBranch,
       }
@@ -133,7 +82,10 @@ export class MergeConflictsDialog extends React.Component<
    */
   private onCancel = async () => {
     const anyResolvedFiles =
-      getResolvedFiles(this.props.workingDirectory).length > 0
+      getResolvedFiles(
+        this.props.workingDirectory,
+        this.props.manualResolutions
+      ).length > 0
     if (!anyResolvedFiles) {
       await this.props.dispatcher.abortMerge(this.props.repository)
       this.props.onDismissed()
@@ -150,7 +102,8 @@ export class MergeConflictsDialog extends React.Component<
 
   private onDismissed = async () => {
     this.props.onDismissed()
-    this.props.dispatcher.setMergeConflictsBannerState({
+    this.props.dispatcher.setBanner({
+      type: BannerType.MergeConflictsFound,
       ourBranch: this.props.ourBranch,
       popup: {
         type: PopupType.MergeConflicts,
@@ -160,7 +113,12 @@ export class MergeConflictsDialog extends React.Component<
       },
     })
     this.props.dispatcher.recordMergeConflictsDialogDismissal()
-    if (getConflictedFiles(this.props.workingDirectory).length > 0) {
+    const anyConflictedFiles =
+      getConflictedFiles(
+        this.props.workingDirectory,
+        this.props.manualResolutions
+      ).length > 0
+    if (anyConflictedFiles) {
       this.props.dispatcher.recordAnyConflictsLeftOnMergeConflictsDialogDismissal()
     }
   }
@@ -189,7 +147,7 @@ export class MergeConflictsDialog extends React.Component<
 
   private renderShellLink(openThisRepositoryInShell: () => void): JSX.Element {
     return (
-      <div className="cli-link">
+      <div>
         <LinkButton onClick={openThisRepositoryInShell}>
           Open in command line,
         </LinkButton>{' '}
@@ -198,136 +156,26 @@ export class MergeConflictsDialog extends React.Component<
     )
   }
 
-  private renderResolvedFile(path: string): JSX.Element {
-    return (
-      <li className="unmerged-file-status-resolved">
-        <Octicon symbol={OcticonSymbol.fileCode} className="file-octicon" />
-        <div className="column-left">
-          <PathText path={path} availableWidth={200} />
-          <div className="file-conflicts-status">No conflicts remaining</div>
-        </div>
-        <div className="green-circle">
-          <Octicon symbol={OcticonSymbol.check} />
-        </div>
-      </li>
-    )
-  }
-
-  private renderConflictedFile(
-    path: string,
-    status: ConflictedFileStatus,
-    onOpenEditorClick: () => void
-  ): JSX.Element | null {
-    let content = null
-    if (isConflictWithMarkers(status)) {
-      const humanReadableConflicts = calculateConflicts(
-        status.conflictMarkerCount
-      )
-      const message =
-        humanReadableConflicts === 1
-          ? `1 conflict`
-          : `${humanReadableConflicts} conflicts`
-
-      const disabled = this.props.resolvedExternalEditor === null
-
-      const tooltip = editorButtonTooltip(this.props.resolvedExternalEditor)
-
-      const onDropdownClick = this.makeDropdownClickHandler(
-        path,
-        this.props.repository.path,
-        this.props.dispatcher
-      )
-
-      content = (
-        <>
-          <div className="column-left">
-            <PathText path={path} availableWidth={200} />
-            <div className="file-conflicts-status">{message}</div>
-          </div>
-          <div className="action-buttons">
-            <Button
-              onClick={onOpenEditorClick}
-              disabled={disabled}
-              tooltip={tooltip}
-              className="small-button button-group-item"
-            >
-              {editorButtonString(this.props.resolvedExternalEditor)}
-            </Button>
-            <Button
-              onClick={onDropdownClick}
-              className="small-button button-group-item arrow-menu"
-            >
-              <Octicon symbol={OcticonSymbol.triangleDown} />
-            </Button>
-          </div>
-        </>
-      )
-    } else {
-      content = (
-        <div>
-          <PathText path={path} availableWidth={400} />
-          <div className="command-line-hint">
-            Use command line to resolve this file
-          </div>
-        </div>
-      )
-    }
-
-    return content !== null ? (
-      <li key={path} className="unmerged-file-status-conflicts">
-        <Octicon symbol={OcticonSymbol.fileCode} className="file-octicon" />
-        {content}
-      </li>
-    ) : null
-  }
-
-  private makeDropdownClickHandler = (
-    relativeFilePath: string,
-    repositoryFilePath: string,
-    dispatcher: Dispatcher
-  ) => {
-    return () => {
-      const absoluteFilePath = join(repositoryFilePath, relativeFilePath)
-      const items: IMenuItem[] = [
-        {
-          label: OpenWithDefaultProgramLabel,
-          action: () => openFile(absoluteFilePath, dispatcher),
-        },
-        {
-          label: RevealInFileManagerLabel,
-          action: () => shell.showItemInFolder(absoluteFilePath),
-        },
-      ]
-      showContextualMenu(items)
-    }
-  }
-
-  private renderUnmergedFile(
-    file: WorkingDirectoryFileChange
-  ): JSX.Element | null {
-    const { status } = file
-    switch (status.kind) {
-      case AppFileStatusKind.Conflicted:
-        if (!hasUnresolvedConflicts(status)) {
-          return this.renderResolvedFile(file.path)
-        }
-
-        return this.renderConflictedFile(file.path, status, () =>
-          this.props.openFileInExternalEditor(
-            join(this.props.repository.path, file.path)
-          )
-        )
-      default:
-        return null
-    }
-  }
-
   private renderUnmergedFiles(
     files: ReadonlyArray<WorkingDirectoryFileChange>
   ) {
     return (
       <ul className="unmerged-file-statuses">
-        {files.map(f => this.renderUnmergedFile(f))}
+        {files.map(f =>
+          isConflictedFile(f.status)
+            ? renderUnmergedFile({
+                path: f.path,
+                status: f.status,
+                resolvedExternalEditor: this.props.resolvedExternalEditor,
+                openFileInExternalEditor: this.props.openFileInExternalEditor,
+                repository: this.props.repository,
+                dispatcher: this.props.dispatcher,
+                manualResolution: this.props.manualResolutions.get(f.path),
+                ourBranch: this.props.ourBranch,
+                theirBranch: this.props.theirBranch,
+              })
+            : null
+        )}
       </ul>
     )
   }
@@ -371,8 +219,10 @@ export class MergeConflictsDialog extends React.Component<
 
   public render() {
     const unmergedFiles = getUnmergedFiles(this.props.workingDirectory)
-    const conflictedFilesCount = getConflictedFiles(this.props.workingDirectory)
-      .length
+    const conflictedFilesCount = getConflictedFiles(
+      this.props.workingDirectory,
+      this.props.manualResolutions
+    ).length
 
     const headerTitle = this.renderHeaderTitle(
       this.props.ourBranch,
