@@ -9,33 +9,21 @@ import { merge } from '../../lib/merge'
 import { caseInsensitiveCompare } from '../../lib/compare'
 import { sanitizedRepositoryName } from '../add-repository/sanitized-repository-name'
 import { Octicon, OcticonSymbol } from '../octicons'
+import {
+  RepositoryPublicationSettings,
+  IDotcomPublicationSettings,
+  PublishSettingsType,
+} from '../../models/publish-settings'
 
 interface IPublishRepositoryProps {
   /** The user to use for publishing. */
   readonly account: Account
 
   /** The settings to use when publishing the repository. */
-  readonly settings: IPublishRepositorySettings
+  readonly settings: RepositoryPublicationSettings
 
   /** The function called when any of the publish settings are changed. */
-  readonly onSettingsChanged: (settings: IPublishRepositorySettings) => void
-}
-
-export interface IPublishRepositorySettings {
-  /** The name to use when publishing the repository. */
-  readonly name: string
-
-  /** The repository's description. */
-  readonly description: string
-
-  /** Should the repository be private? */
-  readonly private: boolean
-
-  /**
-   * The org to which this repository belongs. If null, the repository should be
-   * published as a personal repository.
-   */
-  readonly org: IAPIUser | null
+  readonly onSettingsChanged: (settings: RepositoryPublicationSettings) => void
 }
 
 interface IPublishRepositoryState {
@@ -76,8 +64,8 @@ export class PublishRepository extends React.Component<
     this.setState({ orgs })
   }
 
-  private updateSettings<K extends keyof IPublishRepositorySettings>(
-    subset: Pick<IPublishRepositorySettings, K>
+  private updateSettings<K extends keyof RepositoryPublicationSettings>(
+    subset: Pick<RepositoryPublicationSettings, K>
   ) {
     const existingSettings = this.props.settings
     const newSettings = merge(existingSettings, subset)
@@ -100,14 +88,22 @@ export class PublishRepository extends React.Component<
   }
 
   private onOrgChange = (event: React.FormEvent<HTMLSelectElement>) => {
+    const { settings } = this.props
+    if (settings.kind !== PublishSettingsType.dotcom) {
+      return
+    }
+
     const value = event.currentTarget.value
     const index = parseInt(value, 10)
+    let newSettings: IDotcomPublicationSettings
     if (index < 0 || isNaN(index)) {
-      this.updateSettings({ org: null })
+      newSettings = { ...settings, org: null }
     } else {
       const org = this.state.orgs[index]
-      this.updateSettings({ org })
+      newSettings = { ...settings, org }
     }
+
+    this.props.onSettingsChanged(newSettings)
   }
 
   private renderOrgs(): JSX.Element | null {
@@ -123,7 +119,10 @@ export class PublishRepository extends React.Component<
     )
 
     let selectedIndex = -1
-    const selectedOrg = this.props.settings.org
+    const selectedOrg =
+      this.props.settings.kind === PublishSettingsType.dotcom
+        ? this.props.settings.org
+        : null
     for (const [index, org] of this.state.orgs.entries()) {
       if (selectedOrg && selectedOrg.id === org.id) {
         selectedIndex = index
