@@ -1,8 +1,5 @@
 import * as React from 'react'
-import {
-  PublishRepository,
-  IPublishRepositorySettings,
-} from './publish-repository'
+import { PublishRepository } from './publish-repository'
 import { Dispatcher } from '../dispatcher'
 import { Account } from '../../models/account'
 import { Repository } from '../../models/repository'
@@ -14,10 +11,46 @@ import { getDotComAPIEndpoint } from '../../lib/api'
 import { assertNever, fatalError } from '../../lib/fatal-error'
 import { CallToAction } from '../lib/call-to-action'
 import { getGitDescription } from '../../lib/git'
+import {
+  IDotcomPublicationSettings,
+  IEnterprisePublicationSettings,
+  RepositoryPublicationSettings,
+  PublishSettingsType,
+} from '../../models/publish-settings'
 
 enum PublishTab {
   DotCom = 0,
   Enterprise,
+}
+
+type TabState = IDotcomTabState | IEnterpriseTabState
+
+interface IDotcomTabState {
+  readonly kind: 'dotcom'
+
+  /** The settings for publishing the repository. */
+  readonly settings: IDotcomPublicationSettings
+
+  /**
+   * An error which, if present, is presented to the
+   * user in close proximity to the actions or input fields
+   * related to the current step.
+   */
+  readonly error: Error | null
+}
+
+interface IEnterpriseTabState {
+  readonly kind: 'enterprise'
+
+  /** The settings for publishing the repository. */
+  readonly settings: IEnterprisePublicationSettings
+
+  /**
+   * An error which, if present, is presented to the
+   * user in close proximity to the actions or input fields
+   * related to the current step.
+   */
+  readonly error: Error | null
 }
 
 interface IPublishProps {
@@ -36,23 +69,8 @@ interface IPublishProps {
 interface IPublishState {
   /** The currently selected tab. */
   readonly currentTab: PublishTab
-
-  /** The settings for publishing the repository. */
-  readonly publishSettings: IPublishRepositorySettings
-
-  /**
-   * An error which, if present, is presented to the
-   * user in close proximity to the actions or input fields
-   * related to the current step.
-   */
-  readonly error: Error | null
-
-  /**
-   * The error pertaining to the tab the user is not
-   * actively using. It is swapped when the user changes
-   * the current tab.
-   */
-  readonly inactiveTabError: Error | null
+  readonly dotcomTabState: IDotcomTabState
+  readonly enterpriseTabState: IEnterpriseTabState
 
   /** Is the repository currently being published? */
   readonly publishing: boolean
@@ -72,23 +90,42 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
       startingTab = PublishTab.Enterprise
     }
 
-    const publishSettings = {
+    const publicationSettings = {
       name: props.repository.name,
       description: '',
       private: true,
-      org: null,
+    }
+
+    const dotcomTabState: IDotcomTabState = {
+      kind: 'dotcom',
+      settings: {
+        ...publicationSettings,
+        kind: PublishSettingsType.dotcom,
+        org: null,
+      },
+      error: null,
+    }
+
+    const enterpriseTabState: IEnterpriseTabState = {
+      kind: 'enterprise',
+      settings: {
+        ...publicationSettings,
+        kind: PublishSettingsType.enterprise,
+      },
+      error: null,
     }
 
     this.state = {
       currentTab: startingTab,
-      publishSettings,
-      error: null,
-      inactiveTabError: null,
+      dotcomTabState,
+      enterpriseTabState,
       publishing: false,
     }
   }
 
   public render() {
+    const currentTabState = this.getCurrentTabState()
+
     return (
       <Dialog
         id="publish-repository"
@@ -106,8 +143,8 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
           <span>Enterprise</span>
         </TabBar>
 
-        {this.state.error ? (
-          <DialogError>{this.state.error.message}</DialogError>
+        {currentTabState.error ? (
+          <DialogError>{currentTabState.error.message}</DialogError>
         ) : null}
 
         {this.renderContent()}
@@ -117,14 +154,16 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
   }
 
   public async componentDidMount() {
+    const currentTabState = this.getCurrentTabState()
+
     try {
       const description = await getGitDescription(this.props.repository.path)
       const settings = {
-        ...this.state.publishSettings,
+        ...currentTabState.settings,
         description,
       }
 
-      this.setState({ publishSettings: settings })
+      this.setCurrentTabSettings(settings)
     } catch (error) {
       log.warn(`Couldn't get the repository's description`, error)
     }
@@ -132,12 +171,13 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
 
   private renderContent() {
     const tab = this.state.currentTab
+    const currentTabState = this.getCurrentTabState()
     const account = this.getAccountForTab(tab)
     if (account) {
       return (
         <PublishRepository
           account={account}
-          settings={this.state.publishSettings}
+          settings={currentTabState.settings}
           onSettingsChanged={this.onSettingsChanged}
         />
       )
@@ -146,8 +186,23 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
     }
   }
 
-  private onSettingsChanged = (settings: IPublishRepositorySettings) => {
-    this.setState({ publishSettings: settings })
+  private onSettingsChanged = (settings: RepositoryPublicationSettings) => {
+    let tabState: TabState
+    if (settings.kind === PublishSettingsType.enterprise) {
+      tabState = {
+        kind: 'enterprise',
+        settings: settings,
+        error: this.state.enterpriseTabState.error,
+      }
+    } else {
+      tabState = {
+        kind: 'dotcom',
+        settings: settings,
+        error: this.state.dotcomTabState.error,
+      }
+    }
+
+    this.setTabState(tabState)
   }
 
   private getAccountForTab(tab: PublishTab): Account | null {
@@ -191,7 +246,8 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
   }
 
   private renderFooter() {
-    const disabled = !this.state.publishSettings.name.length
+    const currentTabState = this.getCurrentTabState()
+    const disabled = !currentTabState.settings.name.length
     const tab = this.state.currentTab
     const user = this.getAccountForTab(tab)
     if (user) {
@@ -219,7 +275,10 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
   }
 
   private publishRepository = async () => {
-    this.setState({ error: null, publishing: true })
+    const currentTabState = this.getCurrentTabState()
+
+    this.setCurrentTabError(null)
+    this.setState({ publishing: true })
 
     const tab = this.state.currentTab
     const account = this.getAccountForTab(tab)
@@ -228,7 +287,9 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
       return
     }
 
-    const settings = this.state.publishSettings
+    const settings = currentTabState.settings
+    const org =
+      currentTabState.kind === 'dotcom' ? currentTabState.settings.org : null
 
     try {
       await this.props.dispatcher.publishRepository(
@@ -237,27 +298,56 @@ export class Publish extends React.Component<IPublishProps, IPublishState> {
         settings.description,
         settings.private,
         account,
-        settings.org
+        org
       )
 
       this.props.onDismissed()
     } catch (e) {
-      this.setState({ error: e, publishing: false })
+      this.setCurrentTabError(e)
+      this.setState({ publishing: false })
     }
   }
 
   private onTabClicked = (index: PublishTab) => {
     const isTabChanging = index !== this.state.currentTab
     if (isTabChanging) {
-      // Clear the selected org since dot com and Enterprise will have a different
-      // set of orgs.
-      const settings = { ...this.state.publishSettings, org: null }
-      this.setState({ currentTab: index, publishSettings: settings })
-      // Swap the current stored error from the active tab with the error from
-      // the inactive tab. So that each tab saves and displays their own error.
-      const temporaryError: Error | null = this.state.error
-      this.setState({ error: this.state.inactiveTabError })
-      this.setState({ inactiveTabError: temporaryError })
+      this.setState({ currentTab: index })
     }
+  }
+
+  private getCurrentTabState = () =>
+    this.state.currentTab === PublishTab.DotCom
+      ? this.state.dotcomTabState
+      : this.state.enterpriseTabState
+
+  private setTabState = (state: TabState) => {
+    if (state.kind === 'enterprise') {
+      this.setState({ enterpriseTabState: state })
+    } else {
+      this.setState({ dotcomTabState: state })
+    }
+  }
+
+  private setCurrentTabSettings = (settings: RepositoryPublicationSettings) => {
+    if (settings.kind === PublishSettingsType.enterprise) {
+      const enterpriseTabState = {
+        ...this.state.enterpriseTabState,
+        settings: settings,
+      }
+      this.setTabState(enterpriseTabState)
+    } else {
+      const dotcomTabState = {
+        ...this.state.dotcomTabState,
+        settings: settings,
+      }
+      this.setTabState(dotcomTabState)
+    }
+  }
+
+  private setCurrentTabError = (error: Error | null) => {
+    this.setTabState({
+      ...this.getCurrentTabState(),
+      error: error,
+    })
   }
 }
