@@ -11,10 +11,10 @@ import {
 import { git } from '../../../src/lib/git/core'
 import {
   setupFixtureRepository,
-  mkdirSync,
   setupEmptyRepository,
 } from '../../helpers/repositories'
 import { GitProcess } from 'dugite'
+import { mkdirSync } from '../../helpers/temp'
 
 describe('git/rev-parse', () => {
   let repository: Repository | null = null
