@@ -14,133 +14,214 @@ describe('updateConflictState', () => {
     ['foo', ManualConflictResolutionKind.theirs],
   ])
 
-  it('returns null when no MERGE_HEAD file found', () => {
-    const prevState = createState({
-      conflictState: {
-        currentBranch: 'old-branch',
-        currentTip: 'old-sha',
+  describe('merge conflicts', () => {
+    it('returns null when no MERGE_HEAD file found', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'old-branch',
+          currentTip: 'old-sha',
+          manualResolutions,
+        },
+      })
+      const status = createStatus({ mergeHeadFound: false })
+      const conflictState = updateConflictState(prevState, status, statsStore)
+      expect(conflictState).toBeNull()
+    })
+
+    it('preserves manual resolutions between updates in the same merge', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'old-branch',
+          currentTip: 'old-sha',
+          manualResolutions,
+        },
+      })
+      const status = createStatus({
+        mergeHeadFound: true,
+        currentBranch: 'master',
+        currentTip: 'first-sha',
+      })
+
+      const conflictState = updateConflictState(prevState, status, statsStore)
+
+      expect(conflictState).toEqual({
+        kind: 'merge',
+        currentBranch: 'master',
+        currentTip: 'first-sha',
         manualResolutions,
-      },
-    })
-    const status = createStatus({ mergeHeadFound: false })
-    const conflictState = updateConflictState(prevState, status, statsStore)
-    expect(conflictState).toBeNull()
-  })
-
-  it('preserves manual resolutions between updates in the same merge', () => {
-    const prevState = createState({
-      conflictState: {
-        currentBranch: 'old-branch',
-        currentTip: 'old-sha',
-        manualResolutions,
-      },
-    })
-    const status = createStatus({
-      mergeHeadFound: true,
-      currentBranch: 'master',
-      currentTip: 'first-sha',
+      })
     })
 
-    const conflictState = updateConflictState(prevState, status, statsStore)
+    it('returns null when MERGE_HEAD set but not branch or tip defined', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'old-branch',
+          currentTip: 'old-sha',
+          manualResolutions,
+        },
+      })
+      const status = createStatus({
+        mergeHeadFound: true,
+        currentBranch: undefined,
+        currentTip: undefined,
+      })
 
-    expect(conflictState).toEqual({
-      currentBranch: 'master',
-      currentTip: 'first-sha',
-      manualResolutions,
-    })
-  })
-
-  it('returns null when MERGE_HEAD set but not branch or tip defined', () => {
-    const prevState = createState({
-      conflictState: {
-        currentBranch: 'old-branch',
-        currentTip: 'old-sha',
-        manualResolutions,
-      },
-    })
-    const status = createStatus({
-      mergeHeadFound: true,
-      currentBranch: undefined,
-      currentTip: undefined,
+      const conflictState = updateConflictState(prevState, status, statsStore)
+      expect(conflictState).toBeNull()
     })
 
-    const conflictState = updateConflictState(prevState, status, statsStore)
-    expect(conflictState).toBeNull()
-  })
+    it('returns a value when status has MERGE_HEAD set', () => {
+      const prevState = createState({
+        conflictState: null,
+      })
+      const status = createStatus({
+        mergeHeadFound: true,
+        currentBranch: 'master',
+        currentTip: 'first-sha',
+      })
 
-  it('returns a value when status has MERGE_HEAD set', () => {
-    const prevState = createState({
-      conflictState: null,
-    })
-    const status = createStatus({
-      mergeHeadFound: true,
-      currentBranch: 'master',
-      currentTip: 'first-sha',
-    })
+      const conflictState = updateConflictState(prevState, status, statsStore)
 
-    const conflictState = updateConflictState(prevState, status, statsStore)
-
-    expect(conflictState).toEqual({
-      currentBranch: 'master',
-      currentTip: 'first-sha',
-      manualResolutions: new Map<string, ManualConflictResolution>(),
-    })
-  })
-
-  it('increments abort counter when branch has changed', () => {
-    const prevState = createState({
-      conflictState: {
-        currentBranch: 'old-branch',
-        currentTip: 'old-sha',
+      expect(conflictState).toEqual({
+        kind: 'merge',
+        currentBranch: 'master',
+        currentTip: 'first-sha',
         manualResolutions: new Map<string, ManualConflictResolution>(),
-      },
-    })
-    const status = createStatus({
-      mergeHeadFound: true,
-      currentBranch: 'master',
-      currentTip: 'first-sha',
+      })
     })
 
-    updateConflictState(prevState, status, statsStore)
+    it('increments abort counter when branch has changed', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'old-branch',
+          currentTip: 'old-sha',
+          manualResolutions: new Map<string, ManualConflictResolution>(),
+        },
+      })
+      const status = createStatus({
+        mergeHeadFound: true,
+        currentBranch: 'master',
+        currentTip: 'first-sha',
+      })
 
-    expect(statsStore.recordMergeAbortedAfterConflicts).toHaveBeenCalled()
-  })
+      updateConflictState(prevState, status, statsStore)
 
-  it('increments abort counter when conflict resolved and tip has not changed', () => {
-    const prevState = createState({
-      conflictState: {
+      expect(statsStore.recordMergeAbortedAfterConflicts).toHaveBeenCalled()
+    })
+
+    it('increments abort counter when conflict resolved and tip has not changed', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'master',
+          currentTip: 'old-sha',
+          manualResolutions: new Map<string, ManualConflictResolution>(),
+        },
+      })
+      const status = createStatus({
+        mergeHeadFound: false,
         currentBranch: 'master',
         currentTip: 'old-sha',
-        manualResolutions: new Map<string, ManualConflictResolution>(),
-      },
-    })
-    const status = createStatus({
-      mergeHeadFound: false,
-      currentBranch: 'master',
-      currentTip: 'old-sha',
+      })
+
+      updateConflictState(prevState, status, statsStore)
+
+      expect(statsStore.recordMergeAbortedAfterConflicts).toHaveBeenCalled()
     })
 
-    updateConflictState(prevState, status, statsStore)
+    it('increments success counter when conflict resolved and tip has changed', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'merge',
+          currentBranch: 'master',
+          currentTip: 'old-sha',
+          manualResolutions: new Map<string, ManualConflictResolution>(),
+        },
+      })
+      const status = createStatus({
+        mergeHeadFound: false,
+        currentBranch: 'master',
+        currentTip: 'new-sha',
+      })
 
-    expect(statsStore.recordMergeAbortedAfterConflicts).toHaveBeenCalled()
+      updateConflictState(prevState, status, statsStore)
+
+      expect(statsStore.recordMergeSuccessAfterConflicts).toHaveBeenCalled()
+    })
   })
 
-  it('increments success counter when conflict resolved and tip has changed', () => {
-    const prevState = createState({
-      conflictState: {
+  describe('rebase conflicts', () => {
+    it('returns null when no REBASE_HEAD file found', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'rebase',
+          currentTip: 'old-sha',
+          manualResolutions,
+          targetBranch: 'my-feature-branch',
+          originalBranchTip: 'some-other-sha',
+        },
+      })
+      const status = createStatus({ rebaseContext: null })
+      const conflictState = updateConflictState(prevState, status, statsStore)
+      expect(conflictState).toBeNull()
+    })
+
+    it('returns a value when status has REBASE_HEAD set', () => {
+      const prevState = createState({
+        conflictState: null,
+      })
+      const status = createStatus({
+        rebaseContext: {
+          targetBranch: 'my-feature-branch',
+          originalBranchTip: 'some-other-sha',
+        },
         currentBranch: 'master',
-        currentTip: 'old-sha',
+        currentTip: 'first-sha',
+      })
+
+      const conflictState = updateConflictState(prevState, status, statsStore)
+
+      expect(conflictState).toEqual({
+        kind: 'rebase',
+        currentTip: 'first-sha',
         manualResolutions: new Map<string, ManualConflictResolution>(),
-      },
-    })
-    const status = createStatus({
-      mergeHeadFound: false,
-      currentBranch: 'master',
-      currentTip: 'new-sha',
+        targetBranch: 'my-feature-branch',
+        originalBranchTip: 'some-other-sha',
+      })
     })
 
-    updateConflictState(prevState, status, statsStore)
+    it('preserves manual resolutions when a rebase is detected', () => {
+      const prevState = createState({
+        conflictState: {
+          kind: 'rebase',
+          currentTip: 'old-sha',
+          manualResolutions,
+          targetBranch: 'my-feature-branch',
+          originalBranchTip: 'some-other-sha',
+        },
+      })
+      const status = createStatus({
+        rebaseContext: {
+          targetBranch: 'my-feature-branch',
+          originalBranchTip: 'some-other-sha',
+        },
+        currentBranch: 'master',
+        currentTip: 'first-sha',
+      })
 
-    expect(statsStore.recordMergeSuccessAfterConflicts).toHaveBeenCalled()
+      const conflictState = updateConflictState(prevState, status, statsStore)
+
+      expect(conflictState).toEqual({
+        kind: 'rebase',
+        currentTip: 'first-sha',
+        manualResolutions,
+        targetBranch: 'my-feature-branch',
+        originalBranchTip: 'some-other-sha',
+      })
+    })
   })
 })
